fix(backend): split bot chat repository writes
